+++ app/controllers/user_factions_controller.rb
@@ -1,40 +1,80 @@
 class UserFactionsController < ApplicationController
     def show
-        @user_faction = UserFaction.find(params[:id])
+        if logged_in? && current_user.id === UserFaction.find(params[:id]).user_id || admin?
+            @user_faction = UserFaction.find(params[:id])
+        else 
+            redirect_to login_user_path
+        end
     end 
 
-    def index 
-        @user_factions = UserFaction.all
+    def index
+        if logged_in? 
+            @user_factions = UserFaction.where(user_id: current_user.id) ||""
+        else
+            not_logged_in
+            redirect_to login_path
+        end
     end 
 
     def new
+        if logged_in?
         @user_faction = UserFaction.new
         @factions = Faction.all.map { |x| x.name}
+        else
+            not_logged_in
+            redirect_to login_user_path
+        end
     end 
 
     def create
-        user_faction = UserFaction.new(army_name: params[:user_faction][:army_name], user_id: session[:user_id], faction_id: UserFaction.convert_name_to_id(params[:user_faction][:faction_id]))
+        if logged_in?
+            if UserFaction.find_by(army_name: params[:user_faction][:army_name])
+                army_name_taken
+                redirect_to new_user_faction_path
+            else
 
-        if user_faction.save
-            redirect_to user_faction_path(user_faction)
-        else 
-            save_issue
-            redirect_to new_user_faction_path
+                user_faction = UserFaction.new(army_name: params[:user_faction][:army_name], user_id: session[:user_id], faction_id: Faction.find_by( name: params[:user_faction][:faction_id]).id)
+
+                if user_faction.save
+                    redirect_to user_faction_path(user_faction)
+                else 
+                    save_issue
+                    redirect_to new_user_faction_path
+                end
+            end
+        else
+            not_logged_in
+            redirect_to login_user_path
         end
     end
 
-    def edit 
-        @user_faction = UserFaction.find(params[:id])
+    def edit
+        if logged_in? && current_user.id === UserFaction.find(params[:id]).user_id || admin?
+            @user_faction = UserFaction.find(params[:id])
+        else 
+            not_logged_in
+            redirect_to login_user_path
+        end
     end 
 
-    def update 
-        fuser_faction = UserFaction.find(params[:id])
-        fuser_faction.update(params[:user_faction_params])
-        redirect_to user_faction_path(user_faction)
+    def update
+        if logged_in? && current_user.id === UserFaction.find(params[:id]).user_id || admin?
+            fuser_faction = UserFaction.find(params[:id])
+            fuser_faction.update(params[:user_faction_params])
+            redirect_to user_faction_path(user_faction)
+        else 
+            not_logged_in
+            redirect_to login_user_path
+        end
     end
 
     def destroy
-        user_faction = Faction.find(params[:id])
-        user_faction.destroy
+        if logged_in? && current_user.id === UserFaction.find(params[:id]).user_id || admin?
+            user_faction = Faction.find(params[:id])
+            user_faction.destroy
+        else 
+            not_logged_in
+            redirect_to login_user_path
+        end
     end
 end
